+++ CHANGELOG.md
@@ -3,6 +3,11 @@ This changelog is loose. Versions are not semantic, they are incremental. Splint
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix "Don't know how to create ISeq from: clojure.lang.Symbol" error in `splint.rules.helpers.parse-defn` when trying to parse ill-formed function definitions.
+- "Fix" error messages. Honestly, I'm not great at these so I'm not entirely sure how to best display this stuff.
+
 ## [v1.3.2]
 
 ### Fixed